Remove rkhunter entirely - hopefully temporarily.
It not only poses issues when installing or upgrading because of its rigidity
regarding the "properties" test, but also fails with its default configuration
related to other tests.

It will need to be reworked on, probably restricted to a smaller number of
tests.

The check of /boot/ should probably not be given to rkhunter. Instead, something
like https://github.com/grazzolini/chkboot could be used.

Signed-off-by: KheOps <kheops@ceops.eu>