Add support for wordpress network-mode
Wordpress network-mode allows multiple separate blogs to be served from
the same instance of the software. These can be served as separate
subdirectories or separate subdomains. With the latter option, a plugin
is included to allow separate apex domains to be mapped to each
subdomain.

Wordpress.com is the largest example of wordpress running in network
mode.

This branch also:

- Makes configurable the domain and subdirectory from which wordpress is served
- Refactors the wordpress role
- Includes defaults for all optional parameters
- Includes handlers to ensure these are run when the role is executed in
  isolation
- Updates the role documentation to explain the new options
- Updates wordpress to v4.4.2

Resolves #66