Ensure nginx is running so letsencrypt can work
Handlers only get triggered after all tasks in a play are complete. If a
user runs the nginx role followed by the letsencrypt role in the same
play, letsencrypt fails, as nginx isn't restarted to pick up its new
configuration until after letsencrypt needs it to be available.

This change forces a restart of the nginx service after it has been
configured so that it is available for the letsencrypt certificate
generation process to use.