Support multiple websites
This branch allows Caislean users to specify an arbitrary number of
websites to serve as nginx virtual hosts.

It is possible to use the TLS role to generate a self-signed certificate
that covers more than one website, however users must modify
openssl.conf manually to specify additional domain names as Subject Alt
Names, and this is not yet documented.

Other changes included in this branch:

- Give sites a placeholder index.html file
- Reformat YAML to improve legibility
- Make letsencrypt role support multiple websites
- Use HSTS header only with trusted TLS certificates (closes #56)
- Make git ignore ansible site.retry files

Notes on issue #56:

Some browsers will refuse to allow security exceptions to be made for
self-signed certificates when HSTS is configured. This commit removes
HSTS from the nginx role by default and instead applies it on a
per-domain basis when certificates are obtained from letsencrypt.

The Let's Encrypt certificate authority is trusted by most browsers so
it is safe to use HSTS with its certificates.